+++ certificate-service/src/test/java/com/playdata/certificateservice/CertificateServiceImplApplicationTests.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class CertificateServiceApplicationTests {
+class CertificateServiceImplApplicationTests {
 
     @Test
     void contextLoads() {
