--- vacation-service/src/main/java/com/playdata/vacationservice/vacation/entity/UserStatus.java
@@ -1,7 +1,0 @@
-package com.playdata.vacationservice.vacation.entity;
-
-public enum UserStatus {
-
-    ACTIVE, DELETED
-
-}
